moved an effect to AnimationsEffects folder and added blur to PrecipitationPopUpVC
